+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add a marker **attempt** to indicate the number of attempts to run a test before it passes successfully.
+* Add a method **get_current_attempt** to get the number of the current attempt.
+* Add the ability to run multiple dialog boxes in one test.
+* Fix the problem of tests with dialog boxes freezing.
+* Add autofocus on dialog boxes (on the **Confirm** button or the first item from the list).
 * Remove the **progress** field from the **runstore** database.
 * Add the **hw_id** variable to the **test_stand** field in the database obtained from the stand computer.
 * Add the **location** variables to the **test_stand** field in the database.